Update to cass-operator v1.26.0 (#1581)
* Update to newer cass-operator (v1.26.0-dev), which means updating Kubernetes and controller-runtime dependencies also. This leads to some deprecation / removal changes such as Endpoints -> EndpointSlices (and the associated changes to their behavior), admission.Validator -> CustomValidator for webhooks. Also, just result.Requeue is deprecated and time should be always be defined.

Fixed UpgradeOperator test to pass, although its value is quite limited since it doesn't really test anything.

* Try to update the cass-operator installation for real this time..

* Add delete function for old Endpoints, add e2e tests to verify the endpoints were deleted, return the finalizer check.

* Add hack for kustomize to use Kustomize provided by the GHA runner instead of re-installing it to prevent rate limiter issues. Also, modify upgradeToLatest function in suite_test to avoid finalizer check

* Update to final cass-operator v1.26.0

* update k8ssandra-client to v0.8.1

+++ Makefile
@@ -405,11 +405,16 @@ KUSTOMIZE_INSTALL_SCRIPT ?= "https://raw.githubusercontent.com/kubernetes-sigs/k
 .PHONY: kustomize
 kustomize: $(KUSTOMIZE) ## Download kustomize locally if necessary. If wrong version is installed, it will be removed before downloading.
 $(KUSTOMIZE): $(LOCALBIN)
+ifeq ($(GITHUB_ACTIONS), true)
+	@echo "Running in GitHub Actions, using the kustomize version provided by the runner."
+	test -e $(LOCALBIN)/kustomize || ln -s $(shell which kustomize) $(LOCALBIN)/kustomize
+else
 	@if test -x $(LOCALBIN)/kustomize && ! $(LOCALBIN)/kustomize version | grep -q $(KUSTOMIZE_VERSION); then \
 		echo "$(LOCALBIN)/kustomize version is not expected $(KUSTOMIZE_VERSION). Removing it before installing."; \
 		rm -rf $(LOCALBIN)/kustomize; \
 	fi
 	test -s $(LOCALBIN)/kustomize || { curl -s $(KUSTOMIZE_INSTALL_SCRIPT) | bash -s -- $(subst v,,$(KUSTOMIZE_VERSION)) $(LOCALBIN); }
+endif
 
 .PHONY: controller-gen
 controller-gen: $(CONTROLLER_GEN) ## Download controller-gen locally if necessary. If wrong version is installed, it will be overwritten.
